Module tmc_langs_util::task_executor

Functions

clean

compress_project

compress_tar_for_submitting

extract_and_rewrite_everything

extract_project

extract_project_overwrite

get_exercise_packaging_configuration

is_exercise_root_directory

prepare_solutions
Walks through each path in exercise_map, processing files and copying them into dest_path. Skips hidden directories, directories that contain a .tmcignore file in their root, as well as files matching patterns defined in FILES_TO_SKIP_ALWAYS and directories and files named private. Binary files are copied without extra processing, while text files have solution tags and stubs removed.

prepare_stubs

run_check_code_style

run_tests

scan_exercise