Crate tmc_langs_util

Contains various helpful utilities to be used throughout the tmc-langs project.

Re-exports

pub use error::FileError;

Modules

deserialize

Utility functions for de/serializing data wrapped with serde_path_to_error for better errors.

error

Contains the FileError error type for file operations.

file_util

Various utility functions, primarily wrapping the standard library’s IO and filesystem functions

notification_reporter

Contains an utility for reporting warnings.

parse_util

Contains parse functions that may be convenient for implementing language plugins.

path_util

Utility functions for manipulating paths.

progress_reporter

Utility struct for printing progress reports.

serialize

Utility functions for de/serializing data wrapped with serde_path_to_error for better errors.

Enums

TomlError

Type Aliases

JsonError

YamlError