Crate tmc_langs_util

Contains the task executor

Re-exports

pub use self::error::UtilError;

pub use tmc_langs_framework::anyhow;

Modules

error
Contains the crate error type

file_util
Various utility functions, primarily wrapping the standard library's IO and filesystem functions

progress_reporter

task_executor
Module for calling different tasks of TMC-langs language plug-ins.

Structs

ExerciseDesc
A description of an exercise.

ExercisePackagingConfiguration
Represents configuration based on which submission may be packaged.

RunResult
The result of running an exercise's test suite against a submission.

StyleValidationResult
The result of a style check.

Enums

CommandError
An error caused by a failed attempt to execute an external command.

FileIo
A wrapper for std::io::Error that provides more context for the failed operations.

Language

OutputFormat

RunStatus
The overall status of a test run.

StyleValidationStrategy
Determines how style errors are handled.

TmcError