Module tmc_langs_util::task_executor

Module for calling different tasks of TMC-langs language plug-ins.

Functions

clean
See LanguagePlugin::clean.

compress_project
See LanguagePlugin::compress_project.

compress_tar_for_submitting
Creates a tarball that can be submitted to TMC-sandbox.

extract_project
Finds the correct language plug-in for the given exercise path and calls LanguagePlugin::extract_project, If no language plugin matches, see extract_project_overwrite.

extract_project_overwrite
Extract a given archive file containing a compressed project to a target location. This will overwrite any existing files.

get_exercise_packaging_configuration
See LanguagePlugin::get_exercise_packaging_configuration.

is_exercise_root_directory
Tries to find a language plugin for the path, returning true if one is found.

prepare_solutions
See domain::prepare_solutions.

prepare_stubs
See LanguagePlugin::prepare_stubs.

run_check_code_style
See LanguagePlugin::check_code_style.

run_tests
See LanguagePlugin::run_tests.

scan_exercise
See LanguagePlugin::scan_exercise.